requirements.txt
cloud-sql-python-connector==1.17.0
dnspython==2.7.0
google-auth==2.38.0
httpcore==1.0.7
httpx==0.28.1
Jinja2==3.1.6
jsonpointer==3.0.0
langchain==0.3.20
langchain-core==0.3.43
langchain-openai==0.3.8
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.6
langsmith==0.3.13
livekit==0.21.3
livekit-agents==0.12.17
livekit-api==0.8.2
livekit-plugins-cartesia==0.4.10
livekit-plugins-deepgram==0.7.0
livekit-plugins-openai==0.11.3
livekit-plugins-silero==0.7.5
livekit-protocol==0.9.2
numpy==2.2.4
onnxruntime==1.21.0
openai==1.66.3
orjson==3.10.15
pillow==10.4.0
propcache==0.3.0
protobuf==5.29.3
psutil==5.9
PyMySQL==1.1.1
python-dotenv==1.0.1
python-ulid==3.0.0
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
SQLAlchemy==2.0.38
tiktoken==0.9.0
twilio==9.4.6
types-protobuf==4.25.0.20240417
watchfiles==0.24.0
zstandard==0.23.0